+++ meta/models.py
@@ -5,19 +5,38 @@ import contextlib
 from copy import copy
 
 from django.conf import settings as dj_settings
+from django.db.models import Manager
+from django.utils.functional import cached_property
 
 from . import settings
+from .views import FullUrlMixin
 
 NEED_REQUEST_OBJECT_ERR_MSG = """
 Meta models needs request objects when initializing if sites framework is not used.
 """.strip()
 
 
-class ModelMeta(object):
+class ModelMeta(FullUrlMixin):
     """
     Meta information mixin.
     """
     _metadata = {}
+    """
+    Metadata coniguration dictionary
+
+    `_metadata` dict values can be:
+
+        * name of object method taking the field name as parameter
+        * name of object method taking no parameters
+        * name of object attribute
+        * name of callable taking the field name as parameter
+        * name of callable taking no parameters
+        * literal value
+
+    They are checked in the given order: the first that matches is returned.
+
+    Callable must be available in the module (i.e.: imported if not defined in the module itself)
+    """
     _metadata_default = {
         'title': False,
         'description': False,
@@ -47,6 +66,28 @@ class ModelMeta(object):
         'locale': False,
         'custom_namespace': settings.OG_NAMESPACES,
     }
+    _schema = {}
+    """
+    schema.org properties dictionary
+
+    `_metadata` dict values can be:
+
+        * name of object method taking the field name as parameter
+        * name of object method taking no parameters
+        * name of object attribute
+        * name of callable taking the field name as parameter
+        * name of callable taking no parameters
+        * literal value
+
+    They are checked in the given order: the first that matches is returned.
+
+    Callable must be available in the module (i.e.: imported if not defined in the module itself)
+
+    If the resulting value is a :py:class:`~meta.models.ModelMeta` or :py:class:`~meta.views.Meta` instance
+    its schema is set in the schema.org dataset.
+
+    See :ref:`a sample implementation <schema.model>`.
+    """
 
     def get_meta(self, request=None):
         """
@@ -68,45 +109,70 @@ class ModelMeta(object):
 
     def _get_meta_value(self, field, value):
         """
-        Build the data according to a
+        Build metadata values from :py:attr:`_metadata`
 
         :param field: metadata field name
         :param value: provided value
         :return: data
         """
+
+        def process_value(item):
+            if isinstance(item, Manager):
+                return list(item.all())
+            elif callable(item):
+                try:
+                    return item(field)
+                except TypeError:
+                    return item()
+            else:
+                return item
+
         if value:
             attr = getattr(self, value, False)
             if attr is not False:
-                if callable(attr):
-                    try:
-                        data = attr(field)
-                    except TypeError:
-                        data = attr()
-                else:
-                    data = attr
+                data = process_value(attr)
             else:
-                data = value
+                data = process_value(value)
             return data
 
     def as_meta(self, request=None):
         """
-        Method that generates the Meta object (from django-meta)
+        Populates the :py:class:`~meta.views.Meta` object  with values from :py:attr:`_metadata`
+
+        :param request: optional request object. Used to build the correct URI for linked objects
+        :return: Meta object
         """
         from meta.views import Meta
+
         metadata = self.get_meta(request)
-        meta = Meta(request=request)
+        meta = Meta(request=request, obj=self)
         for field, data in self._retrieve_data(request, metadata):
             setattr(meta, field, data)
         for field in ('og_description', 'twitter_description', 'gplus_description'):
             generaldesc = getattr(meta, 'description', False)
             if not getattr(meta, field, False) and generaldesc:
                 setattr(meta, field, generaldesc)
+        if self._schema:
+            meta.schema = self.schema
         return meta
+
+    @cached_property
+    def schema(self):
+        """
+        Schema.org object description
+
+        :return: dict
+        """
+        schema = {}
+        for field, value in self._schema.items():
+            if value:
+                schema[field] = self._get_meta_value(field, value)
+        return schema
 
     @contextlib.contextmanager
     def _set_request(self, request):
         """
-        Context processor that sets the requst on the current instance
+        Context processor that sets the request on the current instance
         """
         self._request = request
         yield
@@ -172,7 +238,7 @@ class ModelMeta(object):
         """
         Current http protocol
         """
-        return dj_settings.META_SITE_PROTOCOL
+        return self.get_protocol()
 
     def build_absolute_uri(self, url):
         """
@@ -185,15 +251,16 @@ class ModelMeta(object):
         if not dj_settings.META_USE_SITES:
             raise RuntimeError(NEED_REQUEST_OBJECT_ERR_MSG)
 
-        from django.contrib.sites.models import Site
-        s = Site.objects.get_current()
-        meta_protocol = self.get_meta_protocol()
-        if url.startswith('http'):
-            return url
-        if s.domain.find('http') > -1:
-            return '{0}{1}'.format(s.domain, url)  # pragma: no cover
-        else:
-            if url.startswith('/'):
-                return '{0}://{1}{2}'.format(meta_protocol, s.domain, url)
-            else:
-                return '{0}://{1}/{2}'.format(meta_protocol, s.domain, url)
+        return self._get_full_url(url)
+
+    def mainEntityOfPage(self):
+        return {
+            '@type': 'WebPage',
+            '@id': self.build_absolute_uri(self.get_absolute_url())
+        }
+
+    @property
+    def _local_key(self):
+        return '%s:%s:%s' % (
+            self._meta.app_label, self._meta.model_name, self.pk
+        )